move flat rr graph generation stuff out of rr_graph.cpp
