+++ libs/libarchfpga/src/physical_types.cpp
@@ -2,16 +2,8 @@
 #include "arch_types.h"
 #include "vtr_math.h"
 #include "vtr_util.h"
-#include "vtr_log.h"
 
 #include "arch_util.h"
-
-static bool switch_type_is_buffered(SwitchType type);
-static bool switch_type_is_configurable(SwitchType type);
-static e_directionality switch_type_directionality(SwitchType type);
-
-//Ensure the constant has external linkage to avoid linking errors
-constexpr int t_arch_switch_inf::UNDEFINED_FANIN;
 
 /*
  * t_arch_switch_inf
@@ -56,63 +48,7 @@ void t_arch_switch_inf::set_type(SwitchType type_val) {
     type_ = type_val;
 }
 
-/*
- * t_rr_switch_inf
- */
-
-SwitchType t_rr_switch_inf::type() const {
-    return type_;
-}
-
-bool t_rr_switch_inf::buffered() const {
-    return switch_type_is_buffered(type());
-}
-
-bool t_rr_switch_inf::configurable() const {
-    return switch_type_is_configurable(type());
-}
-
-bool t_rr_switch_inf::operator==(const t_rr_switch_inf& other) const {
-    return R == other.R
-           && Cin == other.Cin
-           && Cout == other.Cout
-           && Cinternal == other.Cinternal
-           && Tdel == other.Tdel
-           && mux_trans_size == other.mux_trans_size
-           && buf_size == other.buf_size
-           && power_buffer_type == other.power_buffer_type
-           && power_buffer_size == other.power_buffer_size
-           && intra_tile == other.intra_tile
-           && type() == other.type();
-}
-
-std::size_t t_rr_switch_inf::Hasher::operator()(const t_rr_switch_inf& s) const {
-    std::size_t hash_val = 0;
-
-    auto hash_combine = [&hash_val](auto&& val) {
-        hash_val ^= std::hash<std::decay_t<decltype(val)>>{}(val) + 0x9e3779b9 + (hash_val << 6) + (hash_val >> 2);
-    };
-
-    hash_combine(s.R);
-    hash_combine(s.Cin);
-    hash_combine(s.Cout);
-    hash_combine(s.Cinternal);
-    hash_combine(s.Tdel);
-    hash_combine(s.mux_trans_size);
-    hash_combine(s.buf_size);
-    hash_combine(static_cast<int>(s.power_buffer_type));
-    hash_combine(s.power_buffer_size);
-    hash_combine(s.intra_tile);
-    hash_combine(static_cast<int>(s.type()));
-
-    return hash_val;
-}
-
-void t_rr_switch_inf::set_type(SwitchType type_val) {
-    type_ = type_val;
-}
-
-static bool switch_type_is_buffered(SwitchType type) {
+bool switch_type_is_buffered(SwitchType type) {
     //Muxes and Tristates isolate their input and output into
     //separate DC connected sub-circuits
     return type == SwitchType::MUX
@@ -120,13 +56,13 @@ static bool switch_type_is_buffered(SwitchType type) {
            || type == SwitchType::BUFFER;
 }
 
-static bool switch_type_is_configurable(SwitchType type) {
+bool switch_type_is_configurable(SwitchType type) {
     //Shorts and buffers are non-configurable
     return !(type == SwitchType::SHORT
              || type == SwitchType::BUFFER);
 }
 
-static e_directionality switch_type_directionality(SwitchType type) {
+e_directionality switch_type_directionality(SwitchType type) {
     if (type == SwitchType::SHORT || type == SwitchType::PASS_GATE) {
         //Shorts and pass gates can conduct in either direction
         return e_directionality::BI_DIRECTIONAL;
