+++ libs/librrgraph/src/base/rr_edge.h
@@ -28,3 +28,8 @@ struct t_rr_edge_info {
 };
 
 typedef std::vector<t_rr_edge_info> t_rr_edge_info_set;
+
+inline void uniquify_edges(t_rr_edge_info_set& rr_edges_to_create) {
+    std::stable_sort(rr_edges_to_create.begin(), rr_edges_to_create.end());
+    rr_edges_to_create.erase(std::unique(rr_edges_to_create.begin(), rr_edges_to_create.end()), rr_edges_to_create.end());
+}
