+++ vpr/src/route/rr_graph_generation/rr_graph.h
@@ -41,18 +41,6 @@ void build_tile_rr_graph(RRGraphBuilder& rr_graph_builder,
 
 void free_rr_graph();
 
-t_rr_switch_inf create_rr_switch_from_arch_switch(const t_arch_switch_inf& arch_sw_inf,
-                                                  const float R_minW_nmos,
-                                                  const float R_minW_pmos);
-
-void alloc_and_load_rr_switch_inf(RRGraphBuilder& rr_graph_builder,
-                                  std::vector<std::map<int, int>>& switch_fanin_remap,
-                                  const std::map<int, t_arch_switch_inf>& arch_sw_inf,
-                                  const float R_minW_nmos,
-                                  const float R_minW_pmos,
-                                  const int wire_to_arch_ipin_switch,
-                                  int* wire_to_rr_ipin_switch);
-
 void rr_graph_externals(const std::vector<t_segment_inf>& segment_inf,
                         const std::vector<t_segment_inf>& segment_inf_x,
                         const std::vector<t_segment_inf>& segment_inf_y,
@@ -68,15 +56,6 @@ std::vector<vtr::Matrix<int>> alloc_and_load_actual_fc(const std::vector<t_physi
                                                        const enum e_directionality directionality,
                                                        bool* Fc_clipped,
                                                        bool is_flat);
-
-// Sets the spec for the rr_switch based on the arch switch
-void load_rr_switch_from_arch_switch(RRGraphBuilder& rr_graph_builder,
-                                     const std::map<int, t_arch_switch_inf>& arch_sw_inf,
-                                     int arch_switch_idx,
-                                     int rr_switch_idx,
-                                     int fanin,
-                                     const float R_minW_nmos,
-                                     const float R_minW_pmos);
 
 t_non_configurable_rr_sets identify_non_configurable_rr_sets();
